requirements.txt
anthropic>=0.40.0
google-generativeai>=0.8.0
voyageai>=0.2.3
chromadb>=0.5.0
tree-sitter>=0.23.0
tree-sitter-python>=0.23.0
tree-sitter-javascript>=0.23.0
tree-sitter-typescript>=0.23.0
tree-sitter-go>=0.23.0
tree-sitter-java>=0.23.0
tree-sitter-rust>=0.23.0
rank-bm25>=0.2.2
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-dotenv>=1.0.0
sentence-transformers>=3.0.0